+++ mars/comm/verinfo.h
@@ -2,10 +2,10 @@
 #ifndef Mars_verinfo_h
 #define Mars_verinfo_h
 
-#define MARS_REVISION "66b09bb6"
+#define MARS_REVISION "e072b7c6"
 #define MARS_PATH "master"
 #define MARS_URL ""
-#define MARS_BUILD_TIME "2023-12-04 00:08:13"
+#define MARS_BUILD_TIME "2023-12-04 22:27:01"
 #define MARS_TAG ""
 
 #endif