+++ mars/xlog/flutter_xlog/android/src/main/kotlin/com/tencent/xlog/flutter/FlutterXlogPlugin.kt
@@ -1,7 +1,9 @@
 package com.tencent.xlog.flutter
 
+
 import com.tencent.mars.xlog.Log
 import com.tencent.mars.xlog.Xlog
+import com.tencent.mars.xlog.Xlog.XLogConfig
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
@@ -31,11 +33,19 @@ class FlutterXLogPlugin: FlutterPlugin, MethodCallHandler {
       val logDir = call.argument<String>("logDir")
       val namePrefix = call.argument<String>("namePrefix")
       val cacheDays = call.argument<Int>("cacheDays") ?: 0
+      val pubKey = call.argument<String>("pubKey");
       val consoleLogOpen = call.argument<Boolean>("consoleLogOpen") ?: false
       Log.setLogImp(xlog)
       Log.setLevel(level, false)
       Log.setConsoleLogOpen(consoleLogOpen)
-      xlog.appenderOpen(level, Xlog.AppednerModeAsync, cacheDir, logDir, namePrefix, cacheDays)
+      xlog.appenderOpen(XLogConfig().also {
+        it.level = level
+        it.cachedir = cacheDir
+        it.logdir = logDir
+        it.nameprefix = namePrefix
+        it.cachedays = cacheDays
+        it.pubkey = pubKey
+      })
       result.success(null)
     } else if (call.method == "close") {
       xlog.appenderClose()