+++ mars/xlog/flutter_xlog/ios/Classes/xlog-swift-bridge.h
@@ -2,7 +2,7 @@
 
 @interface XLogBridge: NSObject
 
-- (void)open: (NSUInteger)level cacheDir:(NSString*)cacheDir logDir:(NSString*)logDir prefix:(NSString*)prefix cacheDays:(NSUInteger)cacheDays consoleLogOpen:(BOOL)consoleLogOpen;
+- (void)open: (NSUInteger)level cacheDir:(NSString*)cacheDir logDir:(NSString*)logDir prefix:(NSString*)prefix cacheDays:(NSUInteger)cacheDays pubKey:(NSString*)pubKey, consoleLogOpen:(BOOL)consoleLogOpen;
 
 - (void)close;
 @end