+++ mars/xlog/flutter_xlog/CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Support to use Mars XLog in flutter project.